Stress scripts update.
git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/branches/3.0-proto@3092 7ff1d41c-760d-0410-a7ff-a3a56f310b35

+++ applications/carrot2-webapp/etc/stressgen/stress_redeploy.sh
@@ -1,24 +1,25 @@
 #!/bin/bash
 
-# Source webapp archive
-SOURCE_WAR=/home/dweiss/carrot2/carrot2.sf.proto-3.0/tmp/webapp/carrot2-webapp.war
+#
+# Simple sanity check
+#
 
-# Where to copy it for automatic deployment
-# TODO: we could use tomcat's deploy script here, would take shorter to refresh.
-DEPLOY_WAR=/home/dweiss/Applications/apache-tomcat-6.0.18/webapps/stable.war
+if [ ! -f "env.sh" ]; then
+	echo "Create 'env.sh' based on 'env.sh-template"
+	exit 1
+fi
 
-# Test URI to check if the application is available.
-TEST_URI=http://localhost:8080/stable/
+. env.sh
 
 #
-# Loop deploy/short stress/undeploy
+# Loop between deploy/stress/undeploy
 #
 trap "{ echo; exit 1; }" SIGINT SIGTERM
 while true
 do
     # undeploy
     echo -n "U"
-    rm $DEPLOY_WAR
+    rm -f $DEPLOY_WAR
     while `wget -O /dev/null -q $TEST_URI`
     do
         echo -n "."
@@ -27,7 +28,7 @@ do
     
     # deploy
     echo -n "D"
-    cp $SOURCE_WAR $DEPLOY_WAR
+    ln -s $SOURCE_WAR $DEPLOY_WAR
     while ! `wget -O /dev/null -q $TEST_URI`
     do
         echo -n "."
@@ -39,6 +40,10 @@ do
     # 40 queries are executed at random from the 'queries' file.
     #
     echo -n "S"
+
+    #
+    # Issue 40 queries at random, then quit
+    #
     cat queries | ./stress.rb -t 5 -m 5 -u $TEST_URI --queries 40 2> /dev/null    
 done
